Rename fileutil to ioutil
These functions are not specific to File IO, but are in use for generic
IO. The io_object file seems like a better place.
