+++ include/natalie/file_object.hpp
@@ -14,13 +14,6 @@
 
 namespace Natalie {
 
-namespace fileutil {
-    // Utility Functions Common to File, Dir and Io
-    StringObject *convert_using_to_path(Env *env, Value path);
-    int flags_obj_to_flags(Env *env, IoObject *self, Value flags_obj);
-    mode_t perm_to_mode(Env *env, Value perm);
-}
-
 class FileObject : public IoObject {
 public:
     FileObject()
